gpu: nvgpu: clk: address MISRA 10.3 violations
This fixes a number of miscellaneous MISRA 10.3 violations in
common/pmu/clk/ for assignment of objects of different size or essential
type.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2008771
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/clk/clk.c
@@ -1024,7 +1024,7 @@ int nvgpu_clk_set_boot_fll_clk_tu10x(struct gk20a *g)
 	}
 
 	change_input.pstate_index = 0U;
-	change_input.flags = CTRL_PERF_CHANGE_SEQ_CHANGE_FORCE;
+	change_input.flags = (u32)CTRL_PERF_CHANGE_SEQ_CHANGE_FORCE;
 	change_input.vf_points_cache_counter = 0xFFFFFFFFU;
 
 	status = clk_domain_freq_to_volt(g, gpcclk_domain,
